Added a first try of the test of the Python interface for the structural mechanics part.
However it does not work.
If the notebook is run the kernel will die.
I have traced it back and found out that this happens after one has written something into the map that stores which element has which material.
At first I think that the issue was related to commit `04d6579d0fd` however the fix has not solved the issue, yet.