Added the `addConnectivityType()` function to the Python mesh interface.
The addition of the `MeshAccessor` allowed to manipulate the mesh directly from within Python.
However it was only possible to manipulate _existing_ elements, and not to create new elements of a _different_ not yet existing element type.
In order to offer that the `addConnectivityType()` function of the mesh was also exposed to Python.