This commit fixes several bugs in the the implementation of the Python interface of the `MeshAccessor`, that was introduced by commit `4d0f031e5`.
It was not possible to create a mesh accessor object, because there was no constructor nor function to create it.
Thus a constructor was added.
The second bug was in the wrapper for the `resizeNodes()` function, which actually redirects to the `resizeConnectivity()`.

The last point, was not really an error.
Because of the implementation by functionpointers, the default arguments of the function was lost.
So This was changed.