Updated the Python interface of the material.
The material is parsable, and as such it offers the functions `set()` and `get()` to change and retrieve internal variables.
However, when the parameter changes some internal values might need to be recomputed, which is done by  `Material::updateInternalParameters()` function.
But the `set()` function does not do this.
For that reason the C++ interface of the material offers the `setParam()` function, which basically does this.

This commit now modifies the Python interface of the material and redirects the `set*()` functions, that are defined with by the parameter registry, to the `setParam()` function.
Thus, if a material property is changed by a call to `set*()` the internal values are recomputed.