Implemented a shady way of computing the lumped mass matrix.
For computing the mass associated to a node, the function will iterate through all connections.
It will compute the mass of a beam, which is given as `L * rho * A`, where `L` is the length of the beam, `A` is the cross section area of it and `rho` is the density.
The later two properties are defined through the material object and `L` is computed from the mesh.
The mass is then equally assigned to the two nodes forming the beam.

Computing the rotational inertia is a bit more shady.
The function will compute the "associated volume of a node", which is computed in the same manner as the mass, but the beam volume is given as `A * L`.
It is assumed that this volume is inside a disc (2D) or a sphere (3D).
From this an extension, i.e. radius, can be computed, which then, together with the mass, is used to compute the inertia.

An important note is, that the lumped mass is only stored inside the `StructuralMechanicsModel::mass` member and is not inserted inside the DOFManager.