Updated the computation of the Bernulli beams.
Before the computation of mass of a beam did not consider the cross section area.
It assumed that `StructuralMaterial::rho` had a unit of "mass per unit length.
This commit changes this and now the density is handled with its real units.

The old behaviour can be restored by defined the macro `AKANTU_STRUCTURAL_MECHANICS_OLD_MASS_COMPUTATION`.

It is important that all test are passed.
The reason for this is, that they usually set the corss section to $1$.
And thus the units of `StructuralMaterial::rho` does not has an influence on the result.