Added a check to the `initFullImpl()` function of the `Model` to check the "distribution state" of the model.
This check catches a very exotic bug, that can surfaces if the mesh and the model are not correcly initialized.
If the model, i.e. more precice the DOF manager, is created before the mesh was set to ditributed, then some internal members of it, such as the `residual` or the `data_cache` are created not distributed.
This leads to problem further down, at least in some cases.

This commit catches some form of this bug, by introducing a check inside `initFullImpl()` function of the `akantu::Model` class.
It checks if both the residual and the mesh are either distributed or not distributed.