Updated the `DOFManager`, regarding the new event `MeshIsDistributedEvent`.
The manager listen to that event.
In case it is recived the manager will check if the internal state is consistent with the mesh.
This essentially means if the residual is distributed as well.

In case this is not true an error is generated.
However, it should be possible to reallocate the internals.
meaning to turn the `DOFManager` to a distributed one.
But I do not know enough to do that correctly.

This test used to be inside the `Model::fullInitImpl()`, which was not so good.