Updates the `assembleResidual()` functions of the `StructuralMechanicsModel`.
The commits adds code that ensures that after a call to `assembleResidual()` the matrices `K` and `M` are assembled.
The main point of this commit is, to ensure that after a call to `assumebleResidual()` the `getEnergy()` functions returns a usefull value.
Before they returned zero, because the matrices were not assembled.

This makes the behaviour of the model more consistent to the solid mechanics model.
It is also important for the usage of the structural model inside LibMultiScale.