Updated the implementation of the `StructuralMechanicsModel::getKineticEnergy()` function.
The function now conforms to the computation that are done inside the solid mechanics.
Previously, it was always using the consistent mass for computation the kinetic energy.
However, now, in case it is present it will use the lumped mass.
This is the same way as the solid mechanics works.

There are some small changes in the code, but it is mostly reorderings.
There are two differences.
- In the solid part, nodes whose (lumped) mass is below a certain values are ignored.
	The structural part considers them.
- in the solid part, if neither of the mass matrices are allocated an error is generated.
	The structural part returns zero in that case, which is the original behaviour.
	In addition, also complining to its previous behaviour, it does not syncronize in that case.

Both of the above points could be changed later, to make the model more conforming.