This commit fixes a bug in the `assembleResidual()` function of the `StructuralMechanicsModel`.
In essence the commit amend `29b90b0f220`.
This commit adds also a check for the lumped mass to the `assembleResidual()` function.
Thus it also "ensures" that that matrix exists.

NOTE:
Currently this commit is a bit useless, since the model does not have the ability to assemble the matrix.
However, once it is merged with the developer branch, this should change.