Added some optimization to `SolidMechanicsModel::reassignMaterial()`.
The function now checks if the assignement of a material has changed and only in that case calls `Material::addElements()` and `Material::removeElements()`.
Further the same optimization was applied to the mentioned material functions.
In addition this commit also fixed am error related to the debug system (a missing `AKANTU_DEBUG_OUT()` call was missing in `Material::removeElements()`).